requirements.txt
asyncio
httpx
pyTelegramBotAPI
python-dotenv
aiohttp
motor